HLS Weirdness
So, apparently if your `CABAL_CONFIG` is set to `~/something/config`,
HLS will touch a file at `./~/something/config`. My guess is there's a
directory parser somewhere that doesn't handle `~` correctly. The
effect is that when lsp-haskell starts up HLS, it tries to watch all
the files in the working directory, but it interprets `./~` as `~` and
we get into a _very_ aggressive infinite loop which hangs Emacs until
you can get a `C-g` through.